Change int orders to int orders [8]
Changed the int of orders to an array to track the orders for each menu item.

+++ Program.cpp
@@ -12,7 +12,7 @@ struct menuItemType
 {
     string menuItem;
     double menuPrice;
-    double orders;
+    int orders;
 };
 
 void getData(ifstream& inFile, menuItemType mList[], int listSize);
